Don't crash upgrade script on file not found
Summary:
We now have the same filtering search path issues (essentially we're throwing errors on paths that don't exist) in distillery.

In the meantime, this change is necessary for us to be able to use the upgrade script at all without crashing. We should just skip errors on files that we cannot open the same way we skip files that are generated.

Reviewed By: grievejia

Differential Revision: D17237404

fbshipit-source-id: 4eb2f2d405628e0ad9fe252f3f4c7e5db0e99ede