handle inheritance for attribute models
Summary:
We were previously completely ignoring inheritance for attribute models - this meant that if a subclass attempted to use a sink declared in a parent class, we'd miss it.

Make the global model resolution go through the attribute search and return the parent for the attribute to avoid this issue.

Reviewed By: shannonzhu

Differential Revision: D17242768

fbshipit-source-id: d6c68bc30c23c461d13702edbad1259d220a6ef8