Take environment instead of global resolution in`Check.Signature`
Summary:
The reason for this change is purely technical: we can grab a `dependency_tracked_resolution` from the environment, but we can't grab it from a global resolution.

To reduce the granularity of type checking from per-module to per-function, we'd like to grab different `dependency_tracked_resolution` for different function inside `TypeCheck.run`. Due to the reason mentioned in the previous paragraph, this is never going to happen if `TypeCheck.run` takes `global_resolution` as opposed to `environment`.

Reviewed By: dkgi

Differential Revision: D17274354

fbshipit-source-id: bbb0e3920da7f3c618bfb2589312bc665a3dd16b