Adding actual unannotated globals to the UnannotatedGlobalEnvironment
Summary:
In order to do a clean version of alias resolution, we need to pull  all things that are candidates for becoming aliases into one table in a phase before that.  Luckily, we can do so at the same time as we are registering classes.  These are the eponymous unannotated globals for this phase.  Eventually this will be extended to all kinds of globals that need to end up annotated in the globals table, but for now, it's just the ones for type aliases.

Additionally I added a new variant of dependency that will be used for the separate tracked accesses in the alias phase.

Reviewed By: grievejia

Differential Revision: D17264488

fbshipit-source-id: b71c09c6062a66e51be3b9e55db99f7124f38fbd