Clarifying that first class wins in the tests
Summary:
If you're writing into cached table, the first write still wins in actual shared memory, but the second write actually will overwrite the cache line in the worker you're in.  This means that
```
Worker A:
  add k -> V1
  add k -> V2
  read k => V2

Worker B (or just after a cache eviction):
  read k => V1
```
This meant that in prod we were getting the expected behavior, but in certain tests we could be reading bad values out of the cache.  We now avoid this by using write-through.

Reviewed By: grievejia

Differential Revision: D17265990

fbshipit-source-id: 77dda20b5d8203bff4185d2c1496502b46208b03