Introducing AliasEnvironment
Summary: With the unannotated globals available in the previous phase, we can now use them to implement the new alias resolution algorithm.  The distinction here is that we are no longer allowed to read the Aliases table while calculating aliases in order to preserve the purity of the dependencies.  Luckily, when compared against the benefits of parallelism, this still ends up being a perf win.

Reviewed By: grievejia

Differential Revision: D17266466

fbshipit-source-id: 2d99f1f52fa019d76fb80031e169fad5a491fdea