Move sourcePath into source
Summary:
Goal is to resolve the issue where nonexistent paths get created and errored against in some cases when we set the search path to point to a file inside the project.

Currently our internal source representation doesn't actually hold any absolute path - only a relative path to the project root, which is what we error against in the end.

Source paths that have been renamed therefore have memory of what their real path was, but when we copy the fields over into Source, we lose that.

Reviewed By: grievejia

Differential Revision: D17287311

fbshipit-source-id: 9e8f0277b76ba7341647515c7e51ebde368da6fc