Negation on an integer literal should have type Literal, not int
Summary:
# Background

`Literal[1, 0, -1]` will cause the following analysis errors:
```
test.py:8:10 Invalid type [31]: Expression `typing_extensions.Literal[(1, 0, - 1)]` is not a valid type.
test.py:13:8 Unexpected keyword [28]: Unexpected keyword argument `prop` to call `IntLiteralClazz.__init__`.
```

# Problem

There are two issues that caused this bug, both of which is due to incorrect type inference of the negation of literal integers (i.e. non-variables):

* During type analysis, the definition of `Literal[1, 0, -1]` is not union-ized due to `-1` not being inferred as a literal integer (i.e. kept as `UnaryOperator` or `int` expression type).
* When it is assigned a value (e.g. during init), if the argument is a negative integer, it is inferred as `int` rather than `Literal[-1]`.

# Solution

The change proposed is to check if the operand of the `UnaryOperator` is a constant (i.e. `Integer` type). If so, the type is then inferred to `Literal[-1]` rather than being inferred as `int` (by going through the `int.__neg__(1)` route). This is done as an action in the unary operator production in the parser.

Reviewed By: sinancepel

Differential Revision: D17247966

fbshipit-source-id: 8a9c125f93c8d2548080dd27517e82c59a37ba7a