Introducing `LocalAnnotationMap`
Summary: We were exposing the result type of `ResolutionSharedMemory.get` before. This diff hides that type as well as related operations behind a new module called `LocationAnnotationMap`. The benefit of hiding them is that we get to change how those annotations get stored more freely. For example, this opens up the possibility of only storing the precondition/postcondition and compute the other one on-the-fly, which would cut our type checking shared memory usage in half. Lots of other stuffs need to be done to get there, of course. This is just the starting point.

Reviewed By: dkgi

Differential Revision: D17305264

fbshipit-source-id: aa926a58c3430f4b582a68b9fd73b2756b682c22