lazify class -> callable edges
Summary: Because we have access to the attributes of classes in the type order now, we can look up the `__call__` method on the fly rather than ahead of time.  This is both more principled (as a class with a `__call__` method doesn't actually inherit from Callable[ThatCallable]) and more deterministic, since the attribute creation process used by inferred_callable actually indirectly depends on the class hierarchy being built at the same time.

Reviewed By: dkgi

Differential Revision: D17302440

fbshipit-source-id: 7fe8253e0776b858f150e677e993ae61e429965f