Allow importing from names that has `Any` type
Summary:
We used to error on this:
```
[a.py]
b: Any: ...

[b.py]

import a.b  # error
import a.b.c  # error
from a.b import c  # error
```

Let's change it such that if we try to import from a name that resolves to `Any` type, we don't error on it. After all, the type of `a.b` may be `module`, in which case the error is unjustified.

Reviewed By: dkgi

Differential Revision: D17336596

fbshipit-source-id: ddc8fa80a5d9869706da308f871fc372593abe89