expand override targets 1 level when building dependency graph
Summary: This change brings the construction of the type checking call graph more in line with the existing call graph in `interprocedural/callResolution.ml` - once an override is seen, the logic there will resolve to the base method getting called as well as the overrides of the class that's being called from. Note that it's important to take some care to ensure we don't pick up on unrelated overrides of the base class - an example is provided in the tests.

Reviewed By: fahndrich

Differential Revision: D17322557

fbshipit-source-id: e0f59bb1abd3016cf0eac9589440516b9f6da811