Fix paths for pysa integration tests
Summary:
Missed these when changing how we report errors on files that are included in the search path.

The integration test directory is set up like:

```
django/
   (stubs)
lxml/
   (stubs)
integration_test/
    a.py
    b.py
    .pyre_configuration
```

The pyre configuration sets the search path as: `"search_path": [".."],`.

This means everything will be relative to the parent directory of `integration_test` when it comes to qualifying imports during type checking. You can access things in django stubs with `django.module.Foo`, and you can access source files in integration_test with `integration_test.a.Foo`.

However, when it comes time to error, we now find the true path of the error location and make sure it is relative to the root pyre configuration. In this case, that configuration lives in `integration_test`, so we'll go with that path rather than the relative path we created when taking into account search path renaming.

Reviewed By: grievejia

Differential Revision: D17346121

fbshipit-source-id: 92f6305f354216750219aee887e2e2cdec0d2191