Fix implementation of pyre_extensions.GenericMeta.__getitem__
Summary:
D17192517 didn't actually provide an implementation for the `GenericMeta.__getitem__` function it implemented, thus it returned none, which is not a valid type.
This implements the correct behavior, which is to just indirectly return Generic itself.
I also added some simple tests that will catch this kind of problem.

Reviewed By: shannonzhu

Differential Revision: D17362995

fbshipit-source-id: ace7126afa3b8ba0cef2957bebdef8152348d00b