refactor override computation logic to clarify case c
Summary: While working on the type checking callgraph, I noticed that the way the override computation logic was written for the case where we're creating an override for C.foo when foo is declared in a parent of C was unclear. Make it clear that we always include the base target as well as the override trees for each of C's subclasses.

Reviewed By: dkgi

Differential Revision: D17348528

fbshipit-source-id: b34af40e031d18fe10f8c33ea6931024a12355a9