weaken literals when computing indirect targets
Summary:
String literals behave weirdly in subtyping comparisons - if C subclasses str, C <: Literal['a'] will be false in the subtype check because we can't weaken Literal['a'] to str on the right hand side. However, as the receiver of a method, it can always safely be weakened, so do so.

This fixes a discrepancy between the type checking/existing call graphs.

Reviewed By: dkgi

Differential Revision: D17351320

fbshipit-source-id: 3cf6b91e3673e4405a763bba454add66686f2af3