Bringing AstEnvironment into the same pattern as the other environments
Summary: In order to support the phenomenon of phases depending on not just the immediately previous phase, but potentially all previous phases, we have settled on a pattern where a phase exposes a read only interface, a single writing interface called update, a UpdateResult to encapsulate additions and triggers, and a special depedency key for each phase.  This brings AstEnvironment into this pattern.

Reviewed By: grievejia

Differential Revision: D17367443

fbshipit-source-id: 26d19fbf55c01852da2069f82555971cbc8e3572