Moving Module.from_empty_stub into AstEnvironment.ReadOnly and correcting cache semantics
Summary: This was a pretty awkward fit in Module since you had to pass the most important component in.  Furthermore it belongs to AstEnvironment since it needs to be dependency tracked.  Until we build out the infra for dependency tracked lazy tables we have to avoid the cache for reads with a dependency.

Reviewed By: grievejia

Differential Revision: D17371033

fbshipit-source-id: caf58061f6ea2726fefab4e5da060915e9b452a4