Passing dependency over to GlobalResolution
Summary: Now that we're moving functionality into the individual environments, it is increasingly annoying to have to thread the dependencies over the functions in Environment.resolution_implementation.  In reality, Global resolution is just a container for the environments, along with a dependency to tag all its reads with.  This does that reorganization.

Reviewed By: grievejia

Differential Revision: D17335802

fbshipit-source-id: a40aca5689a47324e2fecceca015894bcb30b466