Error on uninitialized attribute even if init exists
Summary:
We used to ignore checks on all toplevel contents if a constructor existed. (see tests)

Constructor attribute checking used to ignore all superclasses. The one defined in toplevel was much more robust, but the accurate way to go about this is to have the checks be identical in terms of the toplevel verification, and have the constructor check add in one more condition (adding in the info from the post-typecheck resolution of the constructor).

If no constructor exists, do the shell of a check without the extra constructor resolution info. This prevents us from doing double work, having the constuctor/toplevel dependent on one another, or inaccurate responses because the existence of one ignores the contents of the other.

Reviewed By: dkgi

Differential Revision: D17379694

fbshipit-source-id: a555814eb40b575f24c7a6aef5e023e4260faa01