Merge attribute checks between constructor and toplevel
Summary:
Rewriting our logic for attribute-related errors in the toplevel/constructor.

This is messy because:
- We don't want to error on inherited, uninitialized attributes that were not from an abstract or protocol class, because in all other cases, the parent will already be erroring that this attribute is uninitialized
- We need to ensure that the constructor and the toplevel defines are independent of one another (since Jia is working on parallelizing typechecking at the define level, rather than at the module level). However, we want to ensure that we don't miss variables initialized in the init or the toplevel if initializations in both places exist, and also don't want to duplicate work.
- Constructors often have more logic (ie., if statements) that mean we must rely on the post-typecheck resolution to find out what has been initialized or not.

The existing solution was problematic because in the case that an initialization function exists, we are forced to completely ignore the toplevel checks and we wind up with errors that are false positives that are flaky depending on the existence of unrelated pieces of code. We also did a lot of redundant work in the toplevel logic.

The solution I'm working towards looks like --

Constructor:
If this exists, we always perform the typecheck-resolution sensitive checks for uninitialized attributes. We will also check the toplevel using the heuristic of whether a value was given at declaration.

Toplevel:
If the constructor doesn't exist, do a non typecheck-resolution dependent check for uninitialized attributes (even though a resolution exists and we used to use it, this is to be consistent with cases where the constructor exists and prevent dependence of these two defines. In exchange we are giving up a bit of granularity on handling non single-assignment toplevel initialization logic, but that's uncommon)

Any other logic that does not rely on the typecheck-resolution of the constructor should happen here, unconditional on whether a constructor exists or not.

This diff takes the first step of ensuring constructor checks are restricted to typecheck resolution sensitive checks, and ports the remaining checks into the toplevel portion. We used to do separate passes gathering attributes through superclasses for protocols, then abstract classes, but these can be compressed. More diffs to come to whip the rest of this into shape + fix the false positives in distillery generated code.

Reviewed By: dkgi

Differential Revision: D17373443

fbshipit-source-id: bfda1c9cf94f75fab9ff51bbaa6ca133d2d4a386