Eliminating incorrectly implemented callable class decorator functionality
Summary:
If you try to run the given test for this functionality:
```
      import typing
      T = typing.TypeVar("T")
      class synchronize:
        def __call__(
           self,
           coroutine: typing.Callable[..., typing.Coroutine[typing.Any, typing.Any, T]]
        ) -> typing.Callable[..., T]: ...

      synchronize
      async def am_i_async(x: int) -> str:
        return str(x)
      reveal_type(am_i_async)
```
it fails with `TypeError: object() takes no parameters`.  This is because this is trying to call `synchronize()` on the function, which is hitting the default constructor.  Since this is incorrectly implemented, and is blocking our restructuring effort, I'm just killing it for now, and may try tackling it again later if there's demand.

Reviewed By: sinancepel

Differential Revision: D17436297

fbshipit-source-id: 2d9cf3c95969df071f8f8edac40f17b2b5e775fc