Introducing Class Metadata Environment
Summary:
Unfortunately it looks like we're going to need to introduce one more phase, for class metadata.  As it turns out, for perf reasons, the implementation of attributes relies on the successors field from class metadata. Attributes are used for the type order operations needed for global registration.
Luckily, it's fairly straightforward to cut and paste that iteration into its own phase.

Reviewed By: grievejia

Differential Revision: D17414455

fbshipit-source-id: 0fa7e4de6b86e9a4918e359ec58c7d5e8d210c6e