Eliminating incorrectly implemented decorator method implementation
Summary:
Just as in D17436297, there was a feature in unannotated function definition registration that has perf consequences that was not implemented correctly, and thus can be cut.

This is the idea of handling decorators which are themselves methods of a class.  Supporting these requires searching beyond the top level of a source, which is expensive.  However, the implemented version did not get the type of these methods correctly, as it does not drop the leading parameter, and cannot accurately solve against the self parameter in the case of selftypes.

Reviewed By: sinancepel

Differential Revision: D17438965

fbshipit-source-id: ece9b52a4422b9833ae7775a57928231adcc0ffc