Avoid re-computing implicit attributes when performing initialization checks
Summary:
Implicit attribute computation is very expensive. We used to have it in a `O(n^2)` loop -- the computation happens for each attribute in each class, while in principle we can just compute it for each class and re-use the result for each attribute.

A better solution should be to cache those implicit attributes together with explicit ones. But this simple change here is able to buy us several seconds in some of those expensive distillery files.

Reviewed By: shannonzhu

Differential Revision: D17482736

fbshipit-source-id: 3ad35e815d37337099245aa02aaa12590c67fd59