Moving undecorated function registration into ClassHierarchyEnvironmet
Summary:
It turns out that if we're not going to validate the generic parameters in the annotations inside of our undecorated functions (which we can never fully do since that depends on the type order which depends on attributes, which can be decorated with these very functions), we can push undecorated function registration all the way back to the class hierarchy registration phase.

This probably ought to result in a name change for that module, but I haven't come up with one yet.

Reviewed By: grievejia

Differential Revision: D17438369

fbshipit-source-id: ee0b7f5057f91a41ce753a681f2d10956fafc4af