merge obscure models with inferred ones
Summary: An obscure model can be obtained by joining a model for which we declare explicit taint and a subclass which is a stub (and we happen to hit this case in typeshed). When this happens, still compute the model, and join it with the regular obscure taint, taking care to only label the actually obscure taint as such.

Reviewed By: fahndrich

Differential Revision: D17418319

fbshipit-source-id: ce0e368d7a2584ca7bf72ca73dc88a442fcc0012