Do not crash for non-existent site package directory
Summary:
When trying to locate a site package, as long as it exists in _any_ of the site package directories, it should be fine. Our previous logic enforces that the package needs to exist in _all_ of those directories, which is clearly problematic.

It also doesn't make sense to return multiple directories for a given site package: only one of them will be useful to the runtime in the end. This means that `expand_search_path` can simply return a single element instead of a list.

Reviewed By: dkgi

Differential Revision: D17485171

fbshipit-source-id: 6b527ada9e5a95ad26e8efc891ea7124f48efd5c