Move test specification from "old_state & new_state"-style to "old_state & update"-style
Summary:
Previously, we specify the incremental check with an old filesystem state and a new filesystem state. While this model works well for tests where both the base state and the updated state are specified by a mercurial hash, it's not very convenient when we want to compare against a base state and a new state given by some small modifications to the old state --- we'd need to write something like `{ "old_state": OLD_STATE, "new_state": { "kind": "update", "base": OLD_STATE, "modification": MODIFICATION }, ... }`. It's quite annoying that we need to repeat `OLD_STATE` twice, and it's also not efficient to prepare `OLD_STATE` twice during the test.

This diff changes the specification semantics: instead of letting the user specify the new state directly, we instead let the user specify what modifications need to be performed on the old state in order to get the new state (i.e. the specification looks like `{ "old_state": OLD_STATE, "new_state": MODIFICATION, ... }`. This model is strictly more flexible than the previous one when the new state depends on the old state, which is >99% of the time when `pyre_incremental_test` gets invoked.

Core changes are in `specification.py` and `runner.py`. The rest is just tweaking the tests to make them pass.

Reviewed By: dkgi

Differential Revision: D17440227

fbshipit-source-id: 38711a50472fb78a801021128e00dc2d792106da