Put backedge creation back into a localchange
Summary:
Backedges are built up by appending, which is a delete then a write.
Before D17330086 this used to be done in a local change.
This was the right move for edges, which were now parallelized, but didn't realize the implications for backedges.
Hack parallel only reclaims deletions on demand, so our shared memory was like 50+% full of deleted backedges.

This should be huge for both perf and shared memory.

Reviewed By: sinancepel

Differential Revision: D17553473

fbshipit-source-id: a41e0a81db0459ccecde691d8f2d4545bb2f528a