Support specifying updates via patch files
Summary:
Another popular way of updating an existing repository state is to use patch files. These can be generated by `diff`, `git diff`, `hg diff`, etc. The nice thing about patch files is that their size is proportional to the number of *changed lines*, not *changed files*. This makes it very compact if all we want to do is to add/remove one or two lines to a large file.

This diffs adds support to that in our specification: we can specify `kind` of the update as `"patch"`, and provide the patch content as the value to the `"patch"` key.

We use the standard unix `patch` tool to import the specified patches instead of `hg import`. This is intentional: we can't use `hg import` because it's not re-entrant: importing the second time without committing the first import will lead to an error. The same restriction doesn't apply to `patch` -- but the issue with `patch` is that it doesn't know the repository root -- one needs to specify how many layers of path prefix to strip off. Therefore, I choose to leave a "patch_flags" options there to make it possible to fine-tune `patch` behavior if necessary.

Since the `patch` tool typically accepts input from stdin, we also need to extend `Environment.run()` a little bit to make it possible to pass strings to the subprocess' stdin.

Reviewed By: sinancepel

Differential Revision: D17553254

fbshipit-source-id: afdc4e656bc14ba8e42fbea1c42f65fe22aa217d