Support specifying updates via explicit file updates/removals
Summary:
Another popular way of updating an existing repository state is to explicitly specify what files has changed and if so, what's the content after the change. As far as I understand, this is the model that LSP is operating with: whenever the user hits "save" in the editor, the content of the entire file is going to be sent over the wire to the language server, overwriting existing file content.

This diff supports this kind of change by allowing the user to set `kind` of update to `"file"`, and specify what files to update/remove via the `"changes"` and `"removals"` keys. The `changes` value must be a `Dict[str, str]` -- the key represent the file path (relative from old state working directory), and the value represent the file content. The `removals` value must be a list of strings representing paths of the files to be removed. Performing this kind of update is straightforward: we just walk through `changes` and `removals` and update the filesystem accordingly.

Reviewed By: dkgi

Differential Revision: D17554360

fbshipit-source-id: 49be2bbec61cbb8b5c3fe3e94331d6214712c8d2