Support specifying updates via a list of other updates
Summary:
Although limiting ourselves to a single update is good for debugging purpose, sometimes we do want to perform a series of updates in batch for the following reasons:
- A bug might only manifest itself after a series of updates
- Testing individual changes are quite expensive. We want to batch these changes to make the tests more efficient.

This diff adds support to specifying a list of updates via setting `kind` field to `"batch"`. The implementation should be quite straightforward: just recursively go through every single update in the list.

Reviewed By: sinancepel

Differential Revision: D17554682

fbshipit-source-id: 550ed47c826172c96d42af90a3a3b1a549b129ec