Clean up filesystem side-effects when tests are done
Summary:
This diff turns `RepositoryState.prepare()` into a context manager. This allows us to perform some cleanups after the test is done.

For `HgRepositoryState`, which is the only subclass of `RepositoryState` at the moment, we can run `hg whereami` to identify the original commit hash and restore it using `hg update --clean`.

Reviewed By: dkgi

Differential Revision: D17541891

fbshipit-source-id: 35c840025bb1facced9b231997687fb2a7b6dec9