Move pyre-unsafe comment to beneath header
Summary:
This is becoming a bit of a nightmare... we can't actually add `# pyre-unsafe` to the very top of the file, because it might precede a shebang in python scripts.

I've reverted the old change and added some extra logic to ensure we don't undercut any existing pyre ignores at the top of the file, which was introducing errors in the past.

Reviewed By: dkgi

Differential Revision: D17592189

fbshipit-source-id: db13620e8cf64d5a8a51f9efd6256492b5be1f77