Add in Ifs
Summary:
Per T54502976 we need to base the annotated global registration off of the unannotated globals registered in the UnannotatedGlobalEnvironment.  To do this we need to start converging the interfaces here, in part by making sure they have compatible semantics.

This unfortunately for now includes recursing into if statements.  I don't think this should be permitted overall, but for now we'd like to keep this consistent.  The side effect of this is that we start picking up aliases in ifs, but there don't seem to be that many.

Reviewed By: sinancepel

Differential Revision: D17591106

fbshipit-source-id: f1fccbeb3a080e74e359069fec6f2f27da9cbf81