Remove some illegitimate unannotated globals
Summary:
Per T54502976 we need to base the annotated global registration off of the unannotated globals registered in the UnannotatedGlobalEnvironment.  To do this we need to start converging the interfaces here, in part by making sure they have compatible semantics.

The first instance of this is to no longer register foreign assignments as globals, even when those occur at the top level.  This is fine to do since this is a good semantic for aliases as well.

Reviewed By: grievejia

Differential Revision: D17590824

fbshipit-source-id: 27fa3d20608bdc9827e02f4743bbc802e99be6b8