Add missing bases
Summary: When we transition to a unified handling of globals, we're going to start registering *all* class definitions as globals with a meta type.  This includes the "fake" classes we're adding for some built-in types like `Union` or `Literal`.  This is fine, except for right now we're relying on the inferred type of SpecialForm or TypeAlias to give them `__getitem__` functions.  We can instead give them the behavior we want in a more principled fashion by basing them on `typing.Generic[...]`.

Reviewed By: grievejia

Differential Revision: D17591629

fbshipit-source-id: 33fd9274e7d38b7eb10235b926f2bf6abdd2a62e