Record target location in unannotated globals
Summary:
Per T54502976 we need to base the annotated global registration off of the unannotated globals registered in the UnannotatedGlobalEnvironment.  To do this we need to start converging the interfaces here, in part by making sure they have compatible semantics.

Another part of that is to tie the location of the key into the value of the unannotated global, since that's required for the annotated global.  The globals table maps from references to `Annotation.t Node.t` where the location is the location of the key.  This location is used for some error messages.

Reviewed By: grievejia

Differential Revision: D17591010

fbshipit-source-id: 6f4c7be75f65a0d85f65837bd4dfc115195871fe