Test that tuple initialized with generator raises an error.
Summary:
Using `type(some_tuple)(v for v in another_tuple)` or `tuple(v for v in another_tuple)` raises an error about incompatible parameter type. It works fine in the Python interpreter, though, so it should probably typecheck.

The constructor for tuple seems to expect only `List[a]` as a parameter type.

Add a TODO with task number.

Reviewed By: shannonzhu

Differential Revision: D17586263

fbshipit-source-id: 98dd5d5d082b35655d35b4f6ae0e158c77df572d