Support specifying repository state via another repository state plus a list of updates
Summary:
Consider a filesystem state `S` and a series of patches `p0, p1, p2`. With D17554682 we are able to specify a test where `S` is the base and all three patches are batched into the same update.

Let's say that we know testing `S` with updates `p0; p1; p2` fails. To figure out which individual update to blame, we'll want to separate the tests into the following:
1) Base `S` with patch `p0`
2) Base `S+p0` with patch `p1`
3) Base `S+p0+p1` with patch `p2`

We currently don't support specifying tests in (2) and (3) -- this diff fixes that.

Reviewed By: dkgi

Differential Revision: D17575575

fbshipit-source-id: 50623f683754c3e8b4271a9c9f04d6851b55b335