Support specifying repository state via explicitly listing file content
Summary:
So far our incremental tester can only prepare its base state by being in a hg repo and going to a given hg commit. This is fine for large-scale integration tests, but not for small-scale unit tests: Once we find a bug in the integration tests, reduce it to a small repro, and wants to include it in our regression suite, having the option to specify base state via explicitly listing all files becomes very useful.

This diff allows us to set `kind` of the old state to `"file"`, and listing the file contents explicitly in the `"files"` field. The underlying implementation relies on the `FileRepositoryUpdate` introduced in D17554360 to minimize code duplication.

Reviewed By: dkgi

Differential Revision: D17555662

fbshipit-source-id: 5a8f5fd40281127a6b92e2cfd580be3a9e066e70