don't add a receiver argument to staticmethods
Summary: We were previously unconditionally adding a receiver for static methods - this is not the right thing to do for static methods, or instance methods accessed through the class. Fortunately, Pyre provides this information in the callable type directly via a field called "implicit", which if not None is the type of self. In the cases where implicit is None, don't pass a receiver.

Reviewed By: fahndrich

Differential Revision: D17583360

fbshipit-source-id: 012c9a06c66e78d35a672b4b4069f3aec2954e41