Capture all overloads of a define
Summary:
Per T54502976 we need to base the annotated global registration off of the unannotated globals registered in the UnannotatedGlobalEnvironment.  To do this we need to start converging the interfaces here, in part by making sure they have compatible semantics.

Another element of this is the capturing of all overloads of a global function.  We can achieve this while still parallelizing since we know that all instances of a function definition of a given name must occur in a single file.  By capturing all defines, we can also restore the previous functionality of registering the implementation of an overloaded function as a potential decorator.  I think we should actually try to make the whole thing a potential decorator, but that's a story for another time.

Reviewed By: sinancepel

Differential Revision: D17593123

fbshipit-source-id: 0daadcd7aa15f10136136c4c0bd2708d0bdc5348