Add new dependency variants for forthcoming global annotation update
Summary:
We're going to be dependency tracking the production of annotations for globals soon, so we'll need a dependency for that action in all previous phases.

I'm going to be (perhaps a little unfortunately) borrowing GlobalResolution for that purpose as well (except with no globals), since it's power as a environment/dependency pair is useful for that.

Reviewed By: sinancepel

Differential Revision: D17602707

fbshipit-source-id: 798e12e6ca94e87da8e310d0853c6c664802e56e