Add Tuple unannotated global
Summary:
Per T54502976 we need to base the annotated global registration off of the unannotated globals registered in the UnannotatedGlobalEnvironment.  To do this we need to start converging the interfaces here, in part by making sure they have compatible semantics.

The next part of this is registering tuple assignment as a unannotated global.  Globally scoped tuple assignments are valid globals, and as such we need to capture their unannotated forms.  The trick here is that we'd like to keep the property that the keys here are the same as those that will be eventually registered as annotated globals (for invalidation purposes).  However, the natural thing to do would be to key it from Tuple keys -> value, which won't work.  On the other hand, we can't break up the tuple assignment with the information we have now (it could involve class definitions we haven't registered yet).

My solution to this is to instead register the same value for all of the keys, with instructions on how to correctly destructure the corresponding tuple.  These instructions are just an index corresponding to what element we're hoping to associate with the tuple.

This approach has the limitation of not being able to support multi-level tuples (or at least not in the current implementation).  But I think that's a reasonable constraint.

Reviewed By: dkgi

Differential Revision: D17593242

fbshipit-source-id: 13c9e848773dc27840a0fe9c7c2e8e75ec8c1689