Pick up python 2 signatures even if ignore is present
Summary:
This was the root cause of flaky `# pyre-ignores` around python2 style signatures.

It turns out that adding a comment before a comment signature meant we were lexing this into no signature at all.

The error against an invalid type in the python 2 signature therefore disappears when a pyre ignore is added to the line above it, rendering that ignore extraneous.

Note: This wasn't happening in strict mode because we'll throw new errors against missing param/return errors, meaning the ignore still has something to suppress :P

Reviewed By: dkgi

Differential Revision: D17615674

fbshipit-source-id: c51cd044cd5cee05defc2d984dea58daa133c24a