Add decoding for globals
Summary:
I separated this out from the previous diff since it's accessory to the main goals of the environment, but it's important to maintain functionality.  We also need to re-add this stuff for the other new environments.

The only really interesting thing going on here is that we basically have unannotated global environment being a nicely defined source of all possible keys.

Reviewed By: grievejia

Differential Revision: D17614029

fbshipit-source-id: 95f7519e1d0d918ae4f17e9c49d782af8b76c14a