Introducing AnnotatedGlobalEnvironment
Summary:
The long anticipated last environment of environment building is finally here :).
In this diff is a more or less brand new implementation of global resolution that instead of being based off of processing a source, instead processes the revised versions of unannotated global's we've built up in the previous diffs.
This allows this to integrate into the dependency tracked regime without needing to do its own dependency tracking like the implementation in Analysis.Environment does.

This implementation fits the three criteria laid out in T54502976
* No self reads:  This implementation uses a resolution without access to the global table, and relies on UnannotatedGlobalEnvironment to implement the de-duplication semantics, rather than with self-reads
* Consistent Interface:  We now have the same kind of update -> UpdateResult -> Read Only pattern we've come to expect
* Pure environment: This is handled (we directly invalidate based on the request made to us), with the caveat that we have to convert between strings and references.  This should be relatively cheap and is eventually going to be obsoleted by converting all of those primitives from being strings to being references themselves.

Reviewed By: dkgi

Differential Revision: D17614082

fbshipit-source-id: 7e8abf40e334d3872ebeb93d58968577f2ec130f