Temporarily defining Analysis.Environment in terms of AnnotatedGlobalEnvironment
Summary:
For ease of review, I split the rollout of the new environment into two diffs.  This one maintains the public interface of Analysis.Environment and just defines the operations in terms of the new AnnotatedGlobalEnvironment (with some alterations: for example, purge is dead, and the arbitrary dependency tracked update has been merged with register values).  The next one will just kill Analysis.Environment and rewrite all of its callers to instead go directly to AnnotatedGlobalEnvironment.

I split it this way to distinguish functionality changes from naming changes.  This diff is what actually brings the new functionality online, and so is what creates the (small, and actually legit) error differential, as well as a 30s (?!?) perf win.
The next diff is a mechanical find-replace, but is much larger.

Reviewed By: grievejia

Differential Revision: D17579479

fbshipit-source-id: 27256131022783717a57bd8910775f4167035698